pyproject.toml
[project]
name = "bassi"
version = "0.1.0"
description = "Benno's Assistant - Personal AI agent for getting things done"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "rich>=13.9.4",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.7.12",
    "mcp>=1.18.0",
    "claude-agent-sdk>=0.1.4",
    "prompt-toolkit>=3.0.52",
    "msgraph-sdk>=1.46.0",
    "azure-identity>=1.25.1",
    "pillow>=12.0.0",
    "pandas>=2.3.3",
]

[project.scripts]
bassi = "bassi.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-watch>=4.2.0",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 78
target-version = ["py311"]

[tool.ruff]
line-length = 78
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (require API keys)",
    "asyncio: marks tests as async",
]

[dependency-groups]
dev = [
    "pexpect>=4.9.0",
]
